Corrected namespaces for package.
