+++ orcid-model/src/main/java/org/orcid/jaxb/model/common/Source.java
@@ -15,13 +15,13 @@
  * =============================================================================
  */
 //
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.4-2 
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, vJAXB 2.1.10 in JDK 6 
 // See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2014.11.24 at 04:27:39 PM GMT 
+// Generated on: 2012.08.09 at 01:52:56 PM BST 
 //
 
-package org.orcid.jaxb.model.common;
+package org.orcid.jaxb.model.record;
 
 import java.io.Serializable;
 
@@ -39,86 +39,78 @@ import javax.xml.bind.annotation.XmlType;
  * The following schema fragment specifies the expected content contained within
  * this class.
  * 
- * <pre>
- * &lt;complexType>
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;choice>
- *           &lt;element ref="{http://www.orcid.org/ns/common}orcid-id" minOccurs="0"/>
- *           &lt;element ref="{http://www.orcid.org/ns/common}client-id" minOccurs="0"/>
- *         &lt;/choice>
- *         &lt;element ref="{http://www.orcid.org/ns/common}source-name" minOccurs="0"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
  * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = { "clientId", "orcidId", "sourceName" })
-@XmlRootElement(name = "source")
+@XmlType(propOrder = { "sourceOrcid", "sourceClientId", "sourceName" })
+@XmlRootElement(name = "source", namespace = "http://www.orcid.org/ns/common")
 public class Source implements Serializable {
 
-    private final static long serialVersionUID = 1L;
-    @XmlElement(name = "client-id", namespace = "http://www.orcid.org/ns/common")
-    protected ClientId clientId;
-    @XmlElement(name = "orcid-id", namespace = "http://www.orcid.org/ns/common")
-    protected OrcidId orcidId;
-    @XmlElement(name = "source-name", namespace = "http://www.orcid.org/ns/common")
-    protected String sourceName;
-
     /**
-     * Gets the value of the clientId property.
-     * 
-     * @return possible object is {@link ClientId }
      * 
      */
-    public ClientId getClientId() {
-        return clientId;
+    private static final long serialVersionUID = 1L;
+    @XmlElement(name="source-orcid", namespace = "http://www.orcid.org/ns/common")
+    protected SourceOrcid sourceOrcid;
+    @XmlElement(name="source-client-id", namespace="http://www.orcid.org/ns/common")
+    protected SourceClientId sourceClientId;
+    @XmlElement(name="source-name", namespace="http://www.orcid.org/ns/common")
+    protected SourceName sourceName;
+
+    public Source() {
+    }
+
+    public Source(String amenderOrcidPath) {
+        SourceOrcid so = new SourceOrcid();
+        so.setPath(amenderOrcidPath);
+        this.sourceOrcid = so;
     }
 
     /**
-     * Sets the value of the clientId property.
+     * Only use this method if you really need to tell the difference between a
+     * source ORCID and a source client ID. Use {@link #retrieveSourcePath()}
+     * instead.
+     *
+     * @return possible object is {@link SourceOrcid }
+     * 
+     */
+    @Deprecated
+    public SourceOrcid getSourceOrcid() {
+        return sourceOrcid;
+    }
+
+    /**
+     * Sets the value of the sourceOrcid property.
      * 
      * @param value
-     *            allowed object is {@link ClientId }
+     *            allowed object is {@link SourceOrcid }
      * 
      */
-    public void setClientId(ClientId value) {
-        this.clientId = value;
+    public void setSourceOrcid(SourceOrcid value) {
+        this.sourceOrcid = value;
     }
 
     /**
-     * Gets the value of the orcidId property.
-     * 
-     * @return possible object is {@link OrcidId }
-     * 
-     */
-    public OrcidId getOrcidId() {
-        return orcidId;
+     * Only use this method if you really need to tell the difference between a
+     * source ORCID and a source client ID. Use {@link #retrieveSourcePath()}
+     * instead.
+     **/
+    @Deprecated
+    public SourceClientId getSourceClientId() {
+        return sourceClientId;
     }
 
-    /**
-     * Sets the value of the orcidId property.
-     * 
-     * @param value
-     *            allowed object is {@link OrcidId }
-     * 
-     */
-    public void setOrcidId(OrcidId value) {
-        this.orcidId = value;
+    public void setSourceClientId(SourceClientId sourceClientId) {
+        this.sourceClientId = sourceClientId;
     }
 
     /**
      * Gets the value of the sourceName property.
      * 
-     * @return possible object is {@link String }
+     * @return possible object is {@link SourceName }
      * 
      */
-    public String getSourceName() {
+    public SourceName getSourceName() {
         return sourceName;
     }
 
@@ -126,21 +118,52 @@ public class Source implements Serializable {
      * Sets the value of the sourceName property.
      * 
      * @param value
-     *            allowed object is {@link String }
+     *            allowed object is {@link SourceName }
      * 
      */
-    public void setSourceName(String value) {
+    public void setSourceName(SourceName value) {
         this.sourceName = value;
     }
 
     public String retrieveSourcePath() {
-        if (clientId != null) {
-            return clientId.getPath();
+        if (sourceClientId != null) {
+            return sourceClientId.getPath();
         }
-        if (orcidId != null) {
-            return orcidId.getPath();
+        if (sourceOrcid != null) {
+            return sourceOrcid.getPath();
         }
         return null;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((sourceClientId == null) ? 0 : sourceClientId.hashCode());
+        result = prime * result + ((sourceOrcid == null) ? 0 : sourceOrcid.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        Source other = (Source) obj;
+        if (sourceClientId == null) {
+            if (other.sourceClientId != null)
+                return false;
+        } else if (!sourceClientId.equals(other.sourceClientId))
+            return false;
+        if (sourceOrcid == null) {
+            if (other.sourceOrcid != null)
+                return false;
+        } else if (!sourceOrcid.equals(other.sourceOrcid))
+            return false;
+        return true;
     }
 
 }
