+++ orcid-model/src/main/java/org/orcid/jaxb/model/common/Visibility.java
